+++ estampes/visual/molview.py
@@ -382,6 +382,79 @@ class Molecule(Qt3DCore.QEntity):
             self._cam().setViewCenter(abs_pos-loc_pos)
 
 
+class MolWin(Qt3DExtras.Qt3DWindow):
+    """Qt3D Window for the visualization of molecule(s)
+
+    Attributes
+    ----------
+    nmols
+        Number of molecules stored in `atlabs`, `atcrds` and `bonds`.
+    atlabs
+        Atomic labels.
+        If `nmols>1`, list of lists.
+    atcrds
+        3-tuples with atomic coordinates, in Ang.
+        If `nmols>1`, list of lists.
+    bonds
+        2-tuples listing connected atoms.
+        If `nmols>1`, list of lists.
+    col_bond_as_atom
+        If true, bonds are colored based on the connected atoms
+    rad_atom_as_bond
+        If true, atomic radii are set equal to the bonds (tubes).
+    molcols
+        If not `None`, color of the each molecule.
+    """
+    def __init__(self, nmols: int,
+                 atlabs: TypeAtLabM,
+                 atcrds: TypeAtCrdM,
+                 bonds: TypeBondsM,
+                 col_bond_as_atom: bool = False,
+                 rad_atom_as_bond: bool = False,
+                 molcols: tp.Optional[TypeColor] = None):
+        super(MolWin, self).__init__()
+
+        # Camera
+        self.camera().lens().setPerspectiveProjection(45, 16 / 9, 0.1, 1000)
+        self.camera().setPosition(QtGui.QVector3D(0, 1, 40))
+        self.camera().setViewCenter(QtGui.QVector3D(0, 0, 0))
+
+        # For camera controls
+        self.rootEntity = Qt3DCore.QEntity()
+        if nmols == 1:
+            self.mol = Molecule(atlabs, atcrds, bonds, col_bond_as_atom,
+                                rad_atom_as_bond, molcols, self.rootEntity)
+            self.mol.addMouse(self.camera)
+        else:
+            self.mols = []
+            for i in range(nmols):
+                self.mols.append(Molecule(atlabs[i], atcrds[i], bonds[i],
+                                          col_bond_as_atom, rad_atom_as_bond,
+                                          molcols[i], self.rootEntity))
+                self.mols[-1].addMouse(self.camera)
+        self.camController = Qt3DExtras.QOrbitCameraController(self.rootEntity)
+        self.camController.setLinearSpeed(50)
+        self.camController.setLookSpeed(180)
+        self.camController.setCamera(self.camera())
+        self.obj_light = Qt3DCore.QEntity(self.camera())
+        self.camLight = Qt3DRender.QPointLight(self.obj_light)
+        self.cam_tvec = Qt3DCore.QTransform()
+        self.cam_tvec.setTranslation(QtGui.QVector3D(0, 50, 100))
+        self.obj_light.addComponent(self.camLight)
+        self.obj_light.addComponent(self.cam_tvec)
+        # self.camLight.setIntensity(100)
+        # for mol in mols:
+        self.setRootEntity(self.rootEntity)
+
+    def mousePressEvent(self, mouseEvent):
+        if (mouseEvent.button() == QtCore.Qt.RightButton):
+            self.camera().setViewCenter(QtGui.QVector3D(0, 0, 0))
+        # print(mouseEvent.x())
+        # self.camera().setViewCenter(QtGui.QVector3D(mouseEvent.x(),
+        #                                             mouseEvent.y(), 0))
+        super(MolWin, self).mousePressEvent(mouseEvent)
+
+
 # ==============
 # Module Methods
 # ==============